added baseTest
