+++ src/test/java/Pages/BaseTest.java
@@ -1,0 +1,23 @@
+package Pages;
+
+import com.codeborne.selenide.Configuration;
+import com.codeborne.selenide.Selenide;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
+
+public class BaseTest {
+
+
+    @BeforeAll
+    public static void setUp() {
+        WebDriverManager.chromedriver().setup();
+        Configuration.browser = "chrome";
+        Configuration.driverManagerEnabled = true;
+    }
+
+    @AfterEach
+    public void close() {
+        Selenide.closeWebDriver();
+    }
+}
